refactor: Delete CharacterChatAnimation
